refacto: Move view components to common layout folder
