+++ frontend/src/common/layout/show/Show.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ShowBase, ShowProps } from 'react-admin';
 import { ShowActionsWithPermissions } from '@semapps/auth-provider';
-import {ShowView} from '../index';
+import { ShowView } from '../index';
 
 const Show = ({ title, actions, children, ...rest }: ShowProps) => (
   <ShowBase {...rest}>
