+++ frontend/src/resources/Agent/Activity/Event/EventEdit.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { SimpleForm } from "react-admin";
 import { EditToolbarWithPermissions } from "@semapps/auth-provider";
-import { Edit } from "../../../../layout";
+import { Edit } from "../../../../common/layout";
 import EventForm from "./EventForm";
 
 const EventEdit = (props) => (
