+++ frontend/src/resources/Agent/Activity/Project/ProjectList.jsx
@@ -3,7 +3,7 @@ import ProjectFilterSidebar from './ProjectFilterSidebar';
 import { Avatar } from '@mui/material';
 import SettingsIcon from '@mui/icons-material/Settings';
 import SimpleList from "../../../../common/list/SimpleList";
-import { List } from '../../../../layout';
+import { List } from '../../../../common/layout';
 
 const ProjectList = props => (
   <List aside={<ProjectFilterSidebar />} {...props}>
