+++ frontend/src/resources/Agent/Activity/Project/ProjectShow.jsx
@@ -6,7 +6,7 @@ import { ChipList, GridList } from '@semapps/list-components';
 import DescriptionIcon from '@mui/icons-material/Description';
 import { MarkdownField } from '../../../../common/field';
 import { Hero, MainList, SideList } from '../../../../common/list';
-import { Show } from '../../../../layout';
+import { Show } from '../../../../common/layout';
 
 const ProjectShow = props => (
   <Show {...props}>
