+++ frontend/src/resources/Agent/Activity/Task/TaskList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { DateField } from 'react-admin';
 import { Avatar } from '@mui/material';
 import TaskIcon from '@mui/icons-material/PlaylistAddCheck';
-import { List } from '../../../../layout';
+import { List } from '../../../../common/layout';
 import TaskFilterSidebar from './TaskFilterSidebar';
 import SimpleList from "../../../../common/list/SimpleList";
