+++ frontend/src/resources/Agent/Activity/Task/TaskShow.jsx
@@ -5,7 +5,7 @@ import { AvatarWithLabelField, ReferenceArrayField, SeparatedListField } from '@
 import { GridList } from '@semapps/list-components';
 import { MarkdownField } from '../../../../common/field';
 import { Hero, MainList, SideList } from '../../../../common/list';
-import { Show } from '../../../../layout';
+import { Show } from '../../../../common/layout';
 
 const TaskShow = props => (
   <Show {...props}>
