+++ frontend/src/resources/Agent/Actor/Group/GroupList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Avatar } from '@mui/material';
 import GroupIcon from '@mui/icons-material/Group';
-import { List } from '../../../../layout';
+import { List } from '../../../../common/layout';
 import SimpleList from "../../../../common/list/SimpleList";
 
 const GroupList = props => (
