+++ frontend/src/resources/Agent/Actor/Group/GroupShow.jsx
@@ -5,7 +5,7 @@ import { AvatarWithLabelField, QuickAppendReferenceArrayField, ReferenceArrayFie
 import { ChipList, GridList } from '@semapps/list-components';
 import { MarkdownField } from '../../../../common/field';
 import { Hero, MainList, SideList } from '../../../../common/list';
-import { Show } from '../../../../layout';
+import { Show } from '../../../../common/layout';
 
 const GroupShow = props => (
   <Show {...props}>
