+++ frontend/src/resources/Concept/MembershipRole/RoleList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Avatar } from '@mui/material';
 import FavoriteBorderIcon from '@mui/icons-material/Class';
 import SimpleList from "../../../common/list/SimpleList";
-import { List } from '../../../layout';
+import { List } from '../../../common/layout';
 
 const RoleList = props => (
   <List {...props}>
