+++ frontend/src/resources/Concept/Theme/ThemeCreate.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Create } from '../../../layout';
+import { Create } from '../../../common/layout';
 import { LexiconImportForm, fetchWikidata } from "@semapps/interop-components";
 
 const ThemeCreate = (props) => (
