+++ frontend/src/resources/Concept/Theme/ThemeList.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { List } from '../../../layout';
+import { List } from '../../../common/layout';
 import TreeList from '../../../common/list/TreeList';
 
 const ThemeList = props => (
