+++ frontend/src/resources/Concept/Type/TypeEdit.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SimpleForm, TextInput } from 'react-admin';
 import { EditToolbarWithPermissions } from '@semapps/auth-provider';
-import { Edit } from '../../../layout';
+import { Edit } from '../../../common/layout';
 
 export const ThemeEdit = props => (
   <Edit {...props}>
