+++ frontend/src/resources/Concept/Type/TypeList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import StyleIcon from '@mui/icons-material/Style';
 import SimpleList from "../../../common/list/SimpleList";
-import { List } from '../../../layout';
+import { List } from '../../../common/layout';
 
 const TypeList = props => (
   <List {...props}>
