+++ frontend/src/resources/Page/PageCreate.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { SimpleForm, TextInput } from 'react-admin';
-import { Create } from '../../layout';
+import { Create } from '../../common/layout';
 
 const PageCreate = props => (
   <Create {...props}>
