+++ frontend/src/resources/Page/PageList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Avatar } from '@mui/material';
 import DescriptionIcon from '@mui/icons-material/Description';
 import SimpleList from "../../common/list/SimpleList";
-import { List } from '../../layout';
+import { List } from '../../common/layout';
 
 const PageList = props => (
   <List {...props}>
