+++ frontend/src/resources/Resource/Skill/SkillShow.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Grid, Typography } from '@mui/material';
 import { AvatarWithLabelField, QuickAppendReferenceArrayField, ReferenceArrayField } from '@semapps/field-components';
 import { GridList, ChipList } from '@semapps/list-components';
-import { Show } from '../../../layout';
+import { Show } from '../../../common/layout';
 import { SideList } from '../../../common/list';
 
 const SkillShow = props => (
